Rename docs module to docs_util for consistency
Avoid confusion with the important function lettersmith.docs()
