+++ lettersmith.lua
@@ -11,7 +11,7 @@ local concat = lazy.concat
 
 local path = require("lettersmith.path")
 
-local docs = require("lettersmith.docs")
+local docs = require("lettersmith.docs_utils")
 local compare_by_file_path_date = docs.compare_by_file_path_date
 local derive_date = docs.derive_date
