+++ lettersmith/format_date.lua
@@ -12,7 +12,7 @@ Usage:
 
 --]]
 
-local docs = require("lettersmith.docs")
+local docs = require("lettersmith.docs_utils")
 local derive_date = docs.derive_date
 local reformat_yyyy_mm_dd = docs.reformat_yyyy_mm_dd
 local transformer = require("lettersmith.lazy").transformer
